nouveau: remove nouveau_screen reuse (v2)
The screen reuse was introduced to handle GL-VDPAU interop.
Since we have a better solution in place, we can drop this workaround
the respective exported symbol and symbol hacks.

As we follow the same approach for other drivers this will allow us to
remove the VL code from the dri modules and the odd hack that arise around
LLVM use.

Mauro, this should effectively fix the nouveau MT issues so you can drop
the following patches:

 WIP nouveau: add locking (v6)
 nouveau: more locking - make sure that fence work is always done with the push mutex acquired (v2)
 nv30 locking fixes (v2)
 nv30 additional locking fix in context flush
and
 add load_pipe_screen for drm_gralloc (v2)

(v2) Fixes in src/gallium/drivers/nouveau/nouveau_screen.h due to 26ed0f0234
     ("nvc0: increase NOUVEAU_TRANSFER_PUSHBUF_THRESHOLD to 1024 on Kepler+")

Cc: Mauro Rossi <issor.oruam@gmail.com>
Signed-off-by: Emil Velikov <emil.l.velikov@gmail.com>